Add files via upload
This is my first version of a working Pong game to be used for the neurofeedback treatment. There are still many improvements to be made. These improvements include allowing the mouse to move the paddle instead of keyboard, randomizing the computer paddle motion, not allowing the ball to occasionally go through the paddle, recording various elements of the game, and adding cues so the patient knows how to react.